refactor: UI用のディレクトリを用意
