+++ frontend/src/app/page.tsx
@@ -1,5 +1,5 @@
 // 最低限の経由でトップページを表示
-import Home from "./top/page";
+import Home from "./ui/top/page";
 
 export default function Main() {
 	return <Home />;
