+++ frontend/src/app/ui/blog/section/TextSearchSection.tsx
