+++ frontend/src/app/ui/portfolio/page.tsx
@@ -1,4 +1,4 @@
-import Header from "../header";
+import Header from "@/app/components/header";
 
 export default function PortfolioPage() {
   return (
