+++ frontend/src/app/ui/profile/section/ProfileDetailSection.tsx
