+++ frontend/src/app/ui/top/section/WelcomeSection.tsx
